refactor(tests): rename tests/providers/aws/moto/ to tests/providers/aws/mocked/ for provider-neutral convention
Provider-neutral layout lets the k8s provider adopt the same structure
(tests/providers/k8s/mocked/) using kmock-based mocks, without the
directory name implying a specific mocking library.

All import paths, Makefile targets, CI comments, and docs updated.
The moto library imports themselves are untouched.
