+++ tests/providers/aws/mocked/test_asg_price_types.py
