+++ tests/providers/aws/mocked/test_config_driven_provision.py
