+++ tests/providers/aws/mocked/test_cross_cutting.py
