+++ tests/providers/aws/mocked/test_di_wiring.py
