+++ tests/providers/aws/mocked/test_hf_contract.py
