+++ tests/providers/aws/mocked/test_init_discovery_onmoto.py
