+++ tests/providers/aws/mocked/test_launch_template_existing.py
