+++ tests/providers/aws/mocked/test_multi_resource_onmoto.py
