+++ tests/providers/aws/mocked/test_sdk_default_scheduler.py
