+++ tests/providers/aws/mocked/test_template_pipeline.py
